tests/wpt/tests/orientation-event/orientation/t009-manual.https.html

<!DOCTYPE html>
<html>
  <head>
    <title>Rotate the device frame around its x axis</title>
    <meta name=viewport content="width=device-width, maximum-scale=1.0">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <link rel="author" title="Mosquito FP7">
    <link rel="reviewer author" title="Zhiqiang Zhang" href="mailto:zhiqiang.zhang@intel.com"> <!-- 2013-10-12 -->
  </head>
  <body>
    <p>Rotate the device frame around its x axis to run the test.</p>
    <div id="log"></div>
    <script>
      var t1 = async_test("Beta is in [-180, 180)");
      var run = false;
      window.addEventListener("deviceorientation", function(e) {
        if (!run && e.beta !== null) {
          run = true;
          t1.step(function() {
            assert_greater_than_equal(e.beta, -180);
            assert_less_than(e.beta, 180);
          });
          t1.done();
        }
      }, false);
    </script>
  </body>
</html>
